Changed the Save/Copy buttons from toolbar to text field in the chat box.
This fixed the issue where the fading in of the toolbar would stay in the way of a particular line of text that the user would try to select or click (links).

The toolbar implementation was removed and the Save/Copy buttons event dispatching was implemented in the main mxml of the chat box.